refactor: domain karpeta atontzen du
