+++ front/src/domain/basicPage/RegularPageContent.ts
