+++ front/src/domain/hasiera/HasieraContent.ts
