+++ front/src/domain/kafetegia/KafetegiaContent.ts
@@ -1,0 +1,7 @@
+import { MenuKonponentea } from './hooks/useKafetegiaContent';
+
+export interface KafetegiaContent {
+  deskribapena: string;
+  izenburua: string;
+  menua: MenuKonponentea[];
+}
