+++ front/src/domain/kafetegia/mappers/mapAlergenoIdentifikadoreaToIzena.ts
