+++ front/src/domain/kafetegia/models/Alergenoa.ts
