+++ front/src/domain/kafetegia/models/Produktua.ts
