--- front/src/viewQueries/KafetegiaData.ts
@@ -1,7 +1,0 @@
-import { MenuKonponentea } from './useKafetegiaContent';
-
-export interface KafetegiaData {
-  deskribapena: string;
-  izenburua: string;
-  menua: MenuKonponentea[];
-}
