+++ front/src/views/Hasiera/Hasiera.tsx
@@ -5,8 +5,8 @@ import Gezia from '../../assets/gezia.svg';
 import { Container } from '../../components/Container';
 import { Gainburua } from '../../components/Gainburua';
 import { Oina } from '../../components/Oina';
+import { HasieraContent } from '../../domain/hasiera/HasieraContent';
 import * as styles from './Hasiera.module.scss';
-import { HasieraContent } from './HasieraContent';
 
 interface Props {
   content: HasieraContent;
